+++ codegen/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support attributes for inline structs and enums.
 - Add the module `rust::names` with functions, that return various derived names.
 - Support translating links in doc string for Rustdoc.
+- Emit the original doc strings using `#[aldrin(doc = "...")]` for introspection.
 
 ### Changed